refactor(config): update rate limit types in ServerConfig
- Changed rate_limit_per_secs and rate_limit_burst from usize to u64 for improved precision and to accommodate larger values.
- Ensured consistency in data types across configuration settings.

+++ server/src/utils/config.rs
@@ -66,9 +66,9 @@ pub struct ServerConfig {
     #[getset(get = "pub")]
     origins: String,
     #[getset(get = "pub")]
-    rate_limit_per_secs: usize,
+    rate_limit_per_secs: u64,
     #[getset(get = "pub")]
-    rate_limit_burst: usize,
+    rate_limit_burst: u64,
     #[getset(get = "pub")]
     cookie_secret: String,
 }
